build: shorten a bit the makefile using pattern rules
This also fixes the test rule which was incorrect. Use of `\d+`
requires use of PCREs. Extended regular expressions include `+`, but
not `\d`.

Also, normalize a bit the negation of grep result: on a pipeline, the
exit status is the one of the last command. You can negate it by
putting `!` in front of the pipeline.