unicode: fix dangling pointer in sqlite3_unicode_unacc
This function was returning a pointer to a value on the stack,
triggering a GCC warning. This simple (but not entirely correct) fix
for the warning uses a static variable, but this makes the function not
thread safe.

The extension itself only calls the function with a NULL pointer, but
it is marked as SQLITE_EXPORT so could presumably be called by a user.
As this extension is apparently an old import from an unknown
code-generation tool I consider just documenting this limitation to be
sufficient. Users seeking/needing robustness should be using an
alternative.